requirements.txt
openai>=1.1.1
requests>=2.31.0
pyclickup>=0.1.4
Levenshtein>=0.21.1
pypresence>=4.3.0
discord.py>=2.3.2